+++ CHANGELOG.md
@@ -18,3 +18,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   ISO-8601 calendar system, such as 2007-12-03T10:15:30).
 - Added value object `LocationId`.
 - Added `Location`.
+- Added `Locations` - repository of locations.